supervisor/service: get pool txs when wake up
When supervisor processes pool txs, it get all txs in pool before going
to sleep. This behavior causes delay time for all pending txs which come
in to the pool at the time supervisor sleep.

Get the txs when supervisor wakes up, we will reduce the delay time of
pending txs.