p2p/network: use mutex instead of atomic for request nonce
atomic.AddUint64 is using for request nonce counter, it should generate
identical request nonce for each request. However, debugging our
situation show that the request nonce can be duplicated for 2 identical
requests, causing the strange behavior when api query core, an existed
account becomes not found.

It seems to be Go bug itself, but we are not sure. So switching to a
mutex for now.