supervisor/service: fix wrong validators allocation
When calculating validators slice in groupValidators, we ended up making
the validators slice contains empty string, due to:

	make([]string, len(s.Validator))

It causes the slice contains len(s.Validator) empty string "". That code
still pass test, because we only care about the number of validator in
each group. So empty string appears as a valid element.

Just start the slice which cap 0 fix the issue.

	make([]string, 0)