cmd/herserver: remove validator immediately when disconnect
Removing validator when disconnect is done synchronously after a block
time. We want to it to be removed immediately when disconnection occurs,
so moving the check in separated goroutine. It is actually have a 1
second delay, it is good enough, we can increase or remove it later if
it affects.

While at it, also refactoring the code.